feat: add partners and tickets section; remove recaps section and related components
