+++ src/components/sections/partners-and-tickets-section/index.tsx
@@ -1,5 +1,11 @@
+import Partners from './partners'
+
 const PartnersAndTicketsSection = () => {
-  return <section className="bg-[#FCF4F4]" />
+  return (
+    <section className="bg-[#FCF4F4]">
+      <Partners />
+    </section>
+  )
 }
 
 export default PartnersAndTicketsSection
