+++ src/constants/partners.ts
@@ -1,21 +1,29 @@
 import { Partner } from '@/types'
-import gdgLogo from '@public/gdg-logo.png'
+import partner1 from '@public/partners/partner-1.png'
+import partner2 from '@public/partners/partner-2.png'
+import partner3 from '@public/partners/partner-3.png'
+import partner4 from '@public/partners/partner-4.png'
+import partner5 from '@public/partners/partner-5.png'
 
 export const partners: Partner[] = [
   {
-    name: 'Google Developer Groups Ilorin',
-    logo: gdgLogo,
-  },
-  {
-    name: 'Google Developer Groups Kwasu',
-    logo: gdgLogo,
+    name: 'Google Developer Groups Al-Hikmah University',
+    logo: partner1,
   },
   {
     name: 'Google Developer Groups Kwara State Polytechnic',
-    logo: gdgLogo,
+    logo: partner2,
   },
   {
-    name: 'Google Developer Groups Al-Hikmah University',
-    logo: gdgLogo,
+    name: 'Google Developer Groups Kwara State University',
+    logo: partner3,
+  },
+  {
+    name: 'Google Developer Groups Ilorin',
+    logo: partner4,
+  },
+  {
+    name: 'Open Designers',
+    logo: partner5,
   },
 ]
